dom/base/test/test_sandbox_and_document_uri.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test sandbox inheritance and document uri handling</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    add_task(async function() {
      await new Promise((resolve) => {
        window.onmessage = (event) => {
          is(event.data, "done");
          resolve();
        }
      });
    });
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<iframe sandbox="allow-popups allow-scripts" srcdoc=
  "<script>
    window.onmessage = function(event) { parent.postMessage(event.data, '*'); }
    /* Open a cross-origin page */
    window.onload = function() { window.open('http://example.org/tests/dom/base/test/file_sandbox_and_document_uri.html'); }
   </script>"
  ></iframe>
</body>
</html>
